Minor cleanups and improve extension compatibility
